Events handling example

+++ backend/Domain/SalesRevenue.cs
@@ -3,6 +3,6 @@
 public class SalesRevenue
 {
     public Guid Id { get; set; }
-    public double Ammount { get; set; }
+    public decimal Ammount { get; set; }
     public DateTime Date { get; set; }
 }
